fix: One component per file for Panel (#445)
* Rename file

* Rename temp

* Restore original

* Rename file

* Rename temp

* Restore original

* Rename file

* Rename temp

* Restore original

* Rename file

* Rename temp

* Restore original

* Rename file

* Rename temp

* Restore original

* Rename file

* Rename temp

* Restore original

* Rename file

* Rename temp

* Restore original

* Changes to refactor Panel into separate source files

* Updated snapshots

* Remove script file
