+++ src/Panel/_PanelActions.js
@@ -1,0 +1,22 @@
+import classnames from 'classnames';
+import PropTypes from 'prop-types';
+import React from 'react';
+
+const PanelActions = props => {
+    const { children, className, ...rest } = props;
+
+    const panelActionsClasses = classnames(
+        'fd-panel__actions',
+        className
+    );
+
+    return <div {...rest} className={panelActionsClasses}>{children}</div>;
+};
+
+PanelActions.displayName = 'PanelActions';
+
+PanelActions.propTypes = {
+    className: PropTypes.string
+};
+
+export default PanelActions;
