+++ src/index.js
@@ -35,15 +35,8 @@ export { Menu, MenuList, MenuItem, MenuGroup } from './Menu/Menu';
 export { Modal } from './Modal/Modal';
 export { MultiInput } from './MultiInput/MultiInput';
 export { Pagination } from './Pagination/Pagination';
-export { Panel } from './Panel/Panel';
-export { PanelGrid } from './Panel/Panel';
-export { PanelBody } from './Panel/Panel';
-export { PanelHeader } from './Panel/Panel';
-export { PanelHead } from './Panel/Panel';
-export { PanelActions } from './Panel/Panel';
-export { PanelFilters } from './Panel/Panel';
-export { PanelContent } from './Panel/Panel';
-export { PanelFooter } from './Panel/Panel';
+export { default as Panel } from './Panel/Panel';
+export { default as PanelGrid } from './Panel/PanelGrid';
 export { Popover } from './Popover/Popover';
 export { default as ProductTile } from './Tile/ProductTile';
 export { SearchInput } from './SearchInput/SearchInput';
